Simplified design. Had to ditch the USB power option or else ad a charge pump. 2 fewer components. No usb power but USB measurement availible. The current path runs through a few switches so there could be 600mOhm of parasitic resistance. Burden voltage should max out at 60mv @ 100ma assuming poor quality switches. No noise should be coupled into the DUT because the only thing connected is an op amp input and CS resistor.
The Milliohm Meter mode uses a current source of ~10ma for very low resistances.
The Vreg is a 0.4% accuracy source and the resistor under test should be under 1.2ohms to stay within a few percent accuracy.

The probes will double as a 0.4% 1.2v voltage standard for quick testing of multimeters that are cheap.

The battery test LED has been eliminated along with its associated voltage reg and 3 way switch.
Should be in a workable state at this point.